The goal of our project was to create an interactive experience that displayed scenes from Disney's "The Lion King"
We wanted to move the player through three levels of storytelling (with no verbal or direct instruction) while creating an "immersive" environment within our enclosure. To do this we used the code in this repo to light the scene with led's and move servos on command. Different signals like automatic vine openings or blinking lights help guide the player through the enclosure, all the way to the top.